# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/stats/caches": {
            "get": {
                "description": "获取已计算的统计缓存列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计管理"
                ],
                "summary": "获取统计缓存列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "仓库ID（可选，不传则返回所有）",
                        "name": "repo_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "返回数量限制",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/stats/calculate": {
            "post": {
                "description": "异步触发统计计算任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计管理"
                ],
                "summary": "触发统计任务",
                "parameters": [
                    {
                        "description": "统计请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Task"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/stats/commits/count": {
            "get": {
                "description": "统计指定条件下的提交次数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计管理"
                ],
                "summary": "统计提交次数",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "仓库ID",
                        "name": "repo_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分支名称",
                        "name": "branch",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "from",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.CountCommitsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/stats/query": {
            "get": {
                "description": "查询统计计算结果",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计管理"
                ],
                "summary": "查询统计结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "仓库ID",
                        "name": "repo_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分支名称",
                        "name": "branch",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "约束类型",
                        "name": "constraint_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "提交数限制",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.StatsResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ContributorStats": {
            "type": "object",
            "properties": {
                "additions": {
                    "description": "新增行数",
                    "type": "integer"
                },
                "author": {
                    "type": "string"
                },
                "commits": {
                    "type": "integer"
                },
                "deletions": {
                    "description": "删除行数",
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "modifications": {
                    "description": "修改行数 = min(additions, deletions)",
                    "type": "integer"
                },
                "net_additions": {
                    "description": "净增加 = additions - deletions",
                    "type": "integer"
                }
            }
        },
        "models.DateRange": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
...
        "models.Statistics": {
            "type": "object",
            "properties": {
                "by_contributor": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ContributorStats"
                    }
                },
                "summary": {
                    "$ref": "#/definitions/models.StatsSummary"
                }
            }
        },
        "models.StatsConstraint": {
            "type": "object",
            "properties": {
                "from": {
                    "description": "type=date_range时使用",
                    "type": "string"
                },
                "limit": {
                    "description": "type=commit_limit时使用",
                    "type": "integer"
                },
                "to": {
                    "description": "type=date_range时使用",
                    "type": "string"
                },
                "type": {
                    "description": "date_range 或 commit_limit",
                    "type": "string"
                }
            }
        },
        "models.StatsResult": {
            "type": "object",
            "properties": {
                "cache_hit": {
                    "type": "boolean"
                },
                "cached_at": {
                    "type": "string"
                },
                "commit_hash": {
                    "type": "string"
                },
                "statistics": {
                    "$ref": "#/definitions/models.Statistics"
                }
            }
        },
        "models.StatsSummary": {
            "type": "object",
            "properties": {
                "commit_limit": {
                    "type": "integer"
                },
                "date_range": {
                    "$ref": "#/definitions/models.DateRange"
                },
                "total_commits": {
                    "type": "integer"
                },
                "total_contributors": {
                    "type": "integer"
                }
            }
        },
        "models.Task": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration_ms": {
                    "description": "计算字段",
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "parameters": {
                    "description": "JSON string",
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "repo_id": {
                    "type": "integer"
                },
                "result": {
                    "description": "JSON string",
                    "type": "string"
                },
                "retry_count": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "task_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "service.CalculateRequest": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "constraint": {
                    "$ref": "#/definitions/models.StatsConstraint"
                },
                "repo_id": {
                    "type": "integer"
                }
            }
        },
        "service.CountCommitsResponse": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "commit_count": {
                    "type": "integer"
                },
                "from": {
                    "type": "string"
                },
                "repo_id": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                }
            }
        },
...
    }
}